Modify 'format' attribute and handling in Date class
The attribute 'format' in Date class has been changed from a protected instance variable to a public static variable. Accordingly, all method calls to 'format' in the class have been updated to reference the static version. This change allows a uniform date format to be used across all instances of the class.